chore: Move `init_server_config` to `server_config_controller`
