+++ src/rest-api/src/features/init/mod.rs
@@ -3,21 +3,15 @@
 // Copyright: 2023–2025, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
+pub mod errors;
 mod guards;
-mod init_first_account;
-mod init_server_config;
-mod init_workspace;
+mod routes;
 
-use axum::routing::put;
 use axum::routing::MethodRouter;
 
 use crate::AppState;
 
-pub use self::init_first_account::*;
-pub use self::init_server_config::*;
-pub use self::init_workspace::*;
-
-use super::server_config::SERVER_CONFIG_ROUTE;
+pub use self::routes::*;
 
 pub(super) fn router(app_state: AppState) -> axum::Router {
     axum::Router::new()
@@ -27,6 +21,5 @@ pub(super) fn router(app_state: AppState) -> axum::Router {
                 .put(init_first_account_route)
                 .head(is_first_account_created_route),
         )
-        .route(SERVER_CONFIG_ROUTE, put(init_server_config_route))
         .with_state(app_state)
 }
