+++ src/rest-api/src/features/startup_actions/create_service_accounts.rs
@@ -6,7 +6,7 @@
 use service::{server_config::ServerConfigRepository, xmpp::ServerManager};
 use tracing::{debug, info, instrument};
 
-use crate::{features::init::ServerConfigNotInitialized, AppState};
+use crate::{features::init::errors::ServerConfigNotInitialized, AppState};
 
 #[instrument(level = "trace", skip_all, err)]
 pub async fn create_service_accounts(
