+++ src/rest-api/src/features/startup_actions/init_server_config.rs
@@ -8,7 +8,7 @@ use std::sync::Arc;
 use service::{server_config::ServerConfigRepository, xmpp::ServerManager};
 use tracing::{debug, info, instrument};
 
-use crate::{features::init::ServerConfigNotInitialized, AppState};
+use crate::{features::init::errors::ServerConfigNotInitialized, AppState};
 
 #[instrument(level = "trace", skip_all, err)]
 pub async fn init_server_config(
