+++ src/rest-api/src/features/startup_actions/register_oauth2_client.rs
@@ -6,7 +6,7 @@
 use service::server_config::ServerConfigRepository;
 use tracing::{debug, info, instrument};
 
-use crate::{features::init::ServerConfigNotInitialized, AppState};
+use crate::{features::init::errors::ServerConfigNotInitialized, AppState};
 
 #[instrument(level = "trace", skip_all, err)]
 pub async fn register_oauth2_client(
